requirements.txt
streamlit
langchain
langchain-google-genai
redis
yfinance
python-dotenv
langchain-community
beautifulsoup4
openai
langchain-openai